Update library.c
Changed operations on draw_line and draw_pixel methods to use a
custom modulo() method. This method avoids the need to perform an
absolute value operation on the C standard '%' operator. Instead,
a modulo() is performed to keep the line/pixel being drawn within
the bounds of the display.

Also performed formatting on comments and removed the now-unnecessary
abs() method.